+++ internal/app/ui/cpwsarea/wsmap/pmap/psettings/screenshot.go
@@ -147,7 +147,7 @@ func (p *Panel) saveScreenshot(pixels []byte, w, h int) error {
 		directory = cfg.ScreenshotDir
 	}
 
-	dstFilePath := directory + "/StrongDMM-" + time.Now().Format(util.TimeFormat) + ".png"
+	dstFilePath := directory + "/Voidworks-" + time.Now().Format(util.TimeFormat) + ".png"
 
 	if err := saveScreenshotToFile(dstFilePath, pixels, w, h); err != nil {
 		log.Print("unable to save screenshot to file:", err)