+++ internal/env/env.go
@@ -3,9 +3,9 @@ package env
 const Undefined = "undefined"
 
 var (
-	Title    = "StrongDMM - Voidcrew"
-	Version  = "0.5.13"
+	Title    = "Voidworks"
+	Version  = "0.5.14"
 	Revision = Undefined
-	GitHub   = "https://github.com/voidcrew/StrongDMM"
+	GitHub   = "https://github.com/voidcrew/Voidworks"
 	Support  = "https://ko-fi.com/spair"
 )